day 7 - Episode 10 - Added Product Features

+++ app/Http/Controllers/Backend/BrandController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\BrandDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Brand;
+use App\Models\Product;
 use App\Traits\ImageUploadTrait;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Cache;
@@ -103,7 +104,7 @@ class BrandController extends Controller
 
         $brand = Brand::findOrFail($id);
 
-        $imagePath = $this->uploadImage($request, 'logo', 'uploads');
+        $imagePath = $this->updateImage($request, 'logo', 'uploads');
 
         $brand->logo = $imagePath;
         $brand->name = $request->name;
@@ -128,9 +129,9 @@ class BrandController extends Controller
     public function destroy(string $id)
     {
         $brand = Brand::findOrFail($id);
-        // if (Product::where('brand_id', $brand->id)->count() > 0) {
-        //     return response(['status' => 'error', 'message' => 'This brand have products you can\'t delete it.']);
-        // }
+        if (Product::where('brand_id', $brand->id)->count() > 0) {
+            return response(['status' => 'error', 'message' => 'This brand have products you can\'t delete it.']);
+        }
         $this->deleteImage($brand->logo);
         $brand->delete();
